refactor PostType and Taxonomy with base Type
